requirements.txt
fastapi
uvicorn
opencv-python
numpy
torchvision
Pillow
requests